# packages/admin-client/openapi/admin.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlatformRuntimeRequest": {
        "additionalProperties": false,
        "properties": {
          "policy": {
            "$ref": "#/components/schemas/PlatformRuntimePolicy"
          }
        },
        "required": [
          "policy"
        ],
        "title": "PlatformRuntimeRequest",
        "type": "object"
      },
...
      "PlatformRuntimeStateResponse": {
        "properties": {
          "draft_revision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlatformRuntimeRevisionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "latest_published_revision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlatformRuntimeRevisionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "latest_published_revision",
          "draft_revision"
        ],
        "title": "PlatformRuntimeStateResponse",
        "type": "object"
      },
...
      "RuntimeValidationResponse": {
        "properties": {
          "valid": {
            "default": true,
            "title": "Valid",
            "type": "boolean"
          }
        },
        "title": "RuntimeValidationResponse",
        "type": "object"
      },
...
      "TenantRuntimeRequest": {
        "additionalProperties": false,
        "properties": {
          "settings": {
            "$ref": "#/components/schemas/TenantRuntimeOverride"
          }
        },
        "required": [
          "settings"
        ],
        "title": "TenantRuntimeRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AdminToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Agent Platform Admin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/v1/platform/runtime": {
      "get": {
        "operationId": "show_platform_runtime_admin_v1_platform_runtime_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformRuntimeStateResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Show Platform Runtime",
        "tags": [
          "admin:platform-runtime"
        ]
      }
    },
    "/admin/v1/platform/runtime/drafts": {
      "post": {
        "operationId": "create_platform_runtime_draft_admin_v1_platform_runtime_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformRuntimeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformRuntimeRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Create Platform Runtime Draft",
        "tags": [
          "admin:platform-runtime"
        ]
      }
    },
    "/admin/v1/platform/runtime/drafts/{revision_id}": {
      "patch": {
        "operationId": "update_platform_runtime_draft_admin_v1_platform_runtime_drafts__revision_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformRuntimeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformRuntimeRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Update Platform Runtime Draft",
        "tags": [
          "admin:platform-runtime"
        ]
      }
    },
    "/admin/v1/platform/runtime/drafts/{revision_id}/publish": {
      "post": {
        "operationId": "publish_platform_runtime_draft_admin_v1_platform_runtime_drafts__revision_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformRuntimeRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Publish Platform Runtime Draft",
        "tags": [
          "admin:platform-runtime"
        ]
      }
    },
    "/admin/v1/platform/runtime/revisions": {
      "get": {
        "operationId": "list_platform_runtime_revisions_admin_v1_platform_runtime_revisions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlatformRuntimeRevisionResponse"
                  },
                  "title": "Response List Platform Runtime Revisions Admin V1 Platform Runtime Revisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "List Platform Runtime Revisions",
        "tags": [
          "admin:platform-runtime"
        ]
      }
    },
    "/admin/v1/platform/runtime/validate": {
      "post": {
        "operationId": "validate_platform_runtime_admin_v1_platform_runtime_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformRuntimeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Validate Platform Runtime",
        "tags": [
          "admin:platform-runtime"
        ]
      }
    },
    "/admin/v1/tenants": {
      "get": {
        "operationId": "list_tenants_admin_v1_tenants_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TenantResponse"
                  },
                  "title": "Response List Tenants Admin V1 Tenants Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "List Tenants",
        "tags": [
          "admin:tenants"
        ]
      },
      "post": {
        "operationId": "create_tenant_admin_v1_tenants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Create Tenant",
        "tags": [
          "admin:tenants"
        ]
      }
    },
...
    "/admin/v1/tenants/{tenant_id}/runtime": {
      "get": {
        "operationId": "show_tenant_runtime_admin_v1_tenants__tenant_id__runtime_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantRuntimeStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Show Tenant Runtime",
        "tags": [
          "admin:tenant-runtime"
        ]
      }
    },
    "/admin/v1/tenants/{tenant_id}/runtime/drafts": {
      "post": {
        "operationId": "create_tenant_runtime_draft_admin_v1_tenants__tenant_id__runtime_drafts_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantRuntimeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantRuntimeRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Create Tenant Runtime Draft",
        "tags": [
          "admin:tenant-runtime"
        ]
      }
    },
    "/admin/v1/tenants/{tenant_id}/runtime/drafts/{revision_id}": {
      "patch": {
        "operationId": "update_tenant_runtime_draft_admin_v1_tenants__tenant_id__runtime_drafts__revision_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantRuntimeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantRuntimeRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Update Tenant Runtime Draft",
        "tags": [
          "admin:tenant-runtime"
        ]
      }
    },
    "/admin/v1/tenants/{tenant_id}/runtime/drafts/{revision_id}/publish": {
      "post": {
        "operationId": "publish_tenant_runtime_draft_admin_v1_tenants__tenant_id__runtime_drafts__revision_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantRuntimeRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Publish Tenant Runtime Draft",
        "tags": [
          "admin:tenant-runtime"
        ]
      }
    },
    "/admin/v1/tenants/{tenant_id}/runtime/revisions": {
      "get": {
        "operationId": "list_tenant_runtime_revisions_admin_v1_tenants__tenant_id__runtime_revisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TenantRuntimeRevisionResponse"
                  },
                  "title": "Response List Tenant Runtime Revisions Admin V1 Tenants  Tenant Id  Runtime Revisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "List Tenant Runtime Revisions",
        "tags": [
          "admin:tenant-runtime"
        ]
      }
    },
    "/admin/v1/tenants/{tenant_id}/runtime/validate": {
      "post": {
        "operationId": "validate_tenant_runtime_admin_v1_tenants__tenant_id__runtime_validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantRuntimeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Validate Tenant Runtime",
        "tags": [
          "admin:tenant-runtime"
        ]
      }
    },
...
    "/admin/v1/tenants/{tenant_id}/voice-runtime": {
      "get": {
        "operationId": "show_voice_runtime_admin_v1_tenants__tenant_id__voice_runtime_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceRuntimeRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Show Voice Runtime",
        "tags": [
          "admin:tenant-runtime"
        ]
      }
    },
    "/admin/v1/tenants/{tenant_id}/voice-runtime/apply": {
      "post": {
        "operationId": "apply_voice_runtime_admin_v1_tenants__tenant_id__voice_runtime_apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceRuntimeApplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Apply Voice Runtime",
        "tags": [
          "admin:tenant-runtime"
        ]
      }
    },
    "/admin/v1/tenants/{tenant_id}/voice-runtime/plan": {
      "get": {
        "operationId": "plan_voice_runtime_admin_v1_tenants__tenant_id__voice_runtime_plan_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceRuntimePlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Plan Voice Runtime",
        "tags": [
          "admin:tenant-runtime"
        ]
      }
    },
...
  }
}